Fix template API call error parsing
Templates API returns a different error shape than is standard, so we
need to handle that more gracefully here. The client now returns a
useful error message when a bad request is submitted.

Also updated wiremock as I was getting strange random timeouts when
trying to run the test suite.